Add files via upload
probably the repeater coordinates swap approach is too complicated when not using the List Model. This commit is a starting point for future learning goals, but now we'll have to try a different approach to keep from stalling for too long. The new approach will simply be to assign new values ​​for the same coordinates in the onReleased()